Redesigning with musings.groovy
Coalescing everything into an abstract Speaker class.
Might be a good way to reorganize everything. Also
created a Language interface. When fully implemented,
an EnglishSpeaker will implement the language interface
specific to the English language, a Spanish will do the
same for Spanish, and so on. This could also lead to some
interesting possiblities down the road, such as a
BilingualEngSpanSpeaker which inherits from both classes!