Modified IPATernaryTree to reflect new Phon code
Older Phon versions made use of a FeatureComparator class to
construct the IPATernaryTree. The tree, for the purposes of the minimal
pair code in AutoPATT, is used in the search algorithm for constructing
the minimal pairs. The FeatureComparator class has since been removed
from Phon. This commit attempts to reflect this redesign in Phon by
constructing the IPATernaryTree with an IPAElementComparator.